# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Atelier API
  version: "0.1.0"
  description: |
    REST + SSE API for Atelier (canvas-ai) -- an AI-powered UI design generation platform.
    Supports project management, screen generation (including redesign-from-URL and
    image-to-code), component library CRUD, analytics, and feedback collection.
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Optional Bearer token set via ATELIER_API_KEY environment variable.
        When the server has no key configured, all requests are allowed.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    ScreenSummary:
      type: object
      properties:
        id:
          type: string
        prompt:
          type: string
        deviceType:
          type: string
    ScreenDetail:
      type: object
      properties:
        screenId:
          type: string
        html:
          type: string
        prompt:
          type: string
        deviceType:
          type: string
    DesignSystem:
      type: object
      properties:
        colors:
          type: object
          properties:
            primary:
              type: string
            secondary:
              type: string
            tertiary:
              type: string
            background:
              type: string
            neutral:
              type: string
            text:
              type: string
        fonts:
          type: object
          properties:
            headline:
              type: string
            body:
              type: string
        cornerRadius:
          type: string
...
    SSEEvent:
      type: object
      description: >
        Server-Sent Event payload. Each line is prefixed with "data: " and
        contains a JSON object with a "type" field (progress | screen | complete | error).
      properties:
        type:
          type: string
          enum: [progress, screen, complete, error]
      additionalProperties: true
    DeviceType:
      type: string
      enum: [MOBILE, DESKTOP, TABLET, AGNOSTIC]
      default: DESKTOP
paths:
...
  /api/projects/{pid}/screens:
    get:
      tags: [Screens]
      summary: List screens for a project (paginated)
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Paginated screen list
          content:
            application/json:
              schema:
                type: object
                properties:
                  screens:
                    type: array
                    items:
                      $ref: "#/components/schemas/ScreenSummary"
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
                  hasMore:
                    type: boolean
  /api/projects/{pid}/screens/{sid}:
    get:
      tags: [Screens]
      summary: Get a single screen with HTML
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
        - name: sid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Screen details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreenDetail"
        "404":
          description: Screen not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/projects/{pid}/screens/generate:
    post:
      tags: [Screens]
      summary: Generate a new screen (SSE stream)
      description: >
        Streams progress events via SSE. Optionally accepts a sourceUrl for
        redesign mode and a designSystem object to override default tokens.
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
                deviceType:
                  $ref: "#/components/schemas/DeviceType"
                sourceUrl:
                  type: string
                  format: uri
                  description: URL of existing site to redesign (enables redesign mode)
                designSystem:
                  $ref: "#/components/schemas/DesignSystem"
      responses:
        "200":
          description: SSE stream of progress, then complete or error
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/SSEEvent"
        "400":
          description: Missing required field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: SSRF - URL points to private network
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded (30 req/min per IP)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/projects/{pid}/screens/from-image:
    post:
      tags: [Screens]
      summary: Generate a screen from a base64 image
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [imageBase64]
              properties:
                imageBase64:
                  type: string
                  description: Base64-encoded image data
                prompt:
                  type: string
                deviceType:
                  $ref: "#/components/schemas/DeviceType"
      responses:
        "201":
          description: Screen created
          content:
            application/json:
              schema:
                type: object
                properties:
                  screenId:
                    type: string
                  html:
                    type: string
        "400":
          description: Missing required field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/projects/{pid}/screens/{sid}/edit:
    post:
      tags: [Screens]
      summary: Edit an existing screen
      description: >
        Supports full-page and section-level editing. If currentHtml is provided,
        uses the AI refine route; otherwise falls back to the SDK screen edit.
        Optionally accepts sectionSelector and elementSelector for targeted edits.
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
        - name: sid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
                currentHtml:
                  type: string
                  description: Current HTML of the screen (required for AI refine)
                sectionSelector:
                  type: string
                  description: CSS-style selector to target a specific section
                elementSelector:
                  type: string
                elementTag:
                  type: string
                deviceType:
                  $ref: "#/components/schemas/DeviceType"
      responses:
        "200":
          description: Screen edited (full-page or section splice)
          content:
            application/json:
              schema:
                type: object
                properties:
                  screenId:
                    type: string
                  html:
                    type: string
        "201":
          description: Screen edited (SDK fallback path)
          content:
            application/json:
              schema:
                type: object
                properties:
                  screenId:
                    type: string
                  html:
                    type: string
        "400":
          description: Missing required field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Screen not found (when no currentHtml provided)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: AI returned empty or invalid HTML
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/projects/{pid}/screens/{sid}/variants:
    post:
      tags: [Screens]
      summary: Generate a visual variant of a screen
      description: >
        Extracts sections from the provided HTML, generates a variant for each
        section, and reassembles them into a complete page.
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
        - name: sid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt, currentHtml]
              properties:
                prompt:
                  type: string
                currentHtml:
                  type: string
                  description: Full HTML of the current screen
      responses:
        "201":
          description: Array containing the variant
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    screenId:
                      type: string
                    html:
                      type: string
        "400":
          description: Missing or invalid currentHtml, or no sections found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/projects/{pid}/screens/{sid}/export/react:
    get:
      tags: [Screens]
      summary: Export a screen as React code
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
        - name: sid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: React component source
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
        "404":
          description: Screen not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Redesign (multi-page) ──────────────────────────────────────────────
  /api/projects/{pid}/redesign/plan:
    post:
      tags: [Screens]
      summary: Plan a multi-page redesign from a URL (SSE stream)
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                prompt:
                  type: string
      responses:
        "200":
          description: SSE stream ending with a plan object
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/SSEEvent"
        "400":
          description: Missing required field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: SSRF - URL points to private network
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...